refactor: resolve #821 - flatten bottom-right panel to single-level tabs
Replaces two-level tab hierarchy (State Inspector > PALETTE/SPRITE/MOVE,
Buffer Inspector) with single-level tabs: PALETTE, SPRITE, MOVE, BUFFER.
Dissolves StateInspector.vue and BufferInspector.vue wrapper components,
absorbing their content directly into IdeBottomArea.vue.

Closes #821
