+++ src/shared/i18n/locales/ja/ide.json
@@ -389,6 +389,7 @@
     "tabPalette": "PALETTE",
     "tabSprite": "SPRITE",
     "tabMove": "MOVE",
+    "tabBuffer": "BUFFER",
     "tabBg": "BG",
     "palettes": "パレット",
     "backdrop": "バックドロップ",
